Implementation layer controller
