+++ controllers/indexController.js
@@ -1,0 +1,8 @@
+/* GET home page. */
+function getIndex(req, res, next) {
+    res.render('index', { title: 'Express' });
+};
+  
+module.exports = {
+    getIndex
+};
