+++ data/users.json
@@ -1,1 +1,1 @@
-[{"nome":"Rubi Sorrentino","idade":18,"uf":"SP","id":"7eb0582b-8e6c-45d9-81f7-4fd566e991ad"},{"nome":"Nicolas Ferreira Almeida","idade":45,"uf":"SP","id":"af497040-0544-44da-b751-4017a58c6430"},{"nome":"Tarcio Rodrigo A Santos","idade":45,"uf":"SP","email":"tarciorodrigoa@gmail.com","id":"bbbedc54-3584-46eb-966b-5561dc11cb41"}]
+[{"nome":"Rubi Sorrentino","idade":18,"uf":"SP","id":"7eb0582b-8e6c-45d9-81f7-4fd566e991ad"},{"nome":"Nicolas Ferreira Almeida","idade":45,"uf":"SP","id":"af497040-0544-44da-b751-4017a58c6430"},{"nome":"Tarcio Rodrigo Almeida dos Santos","idade":45,"uf":"SP","email":"tarciorodrigoa@gmail.com","id":"bbbedc54-3584-46eb-966b-5561dc11cb41"},{"nome":"Nataly Ferreira","idade":23,"uf":"SP","id":"7c4c351f-3099-47ea-8a12-a61839261ba9"}]
