+++ routes/index.js
@@ -1,9 +1,8 @@
 const express = require('express');
 const router = express.Router();
+const indexController = require("../controllers/indexController");
 
 /* GET home page. */
-router.get('/', (req, res, next) => {
-  res.render('index', { title: 'Express' });
-});
+router.get('/', indexController.getIndex);
 
 module.exports = router;
